+++ noodles-vcf/examples/vcf_query_async.rs
@@ -24,8 +24,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .map(bgzf::AsyncReader::new)
         .map(vcf::AsyncReader::new)?;
 
+    let header = reader.read_header().await?.parse()?;
+
     let index = tabix::r#async::read(src.with_extension("gz.tbi")).await?;
-    let mut query = reader.query(&index, &region)?;
+    let mut query = reader.query(&header, &index, &region)?;
 
     while let Some(record) = query.try_next().await? {
         println!("{}", record);